separate Sakamoto and Sakamoto.Packet
