+++ Sakamoto.Packet/Sakamoto.Packet.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\osu.Shared\osu.Shared\osu.Shared.csproj" />
+  </ItemGroup>
+
+</Project>
