+++ Sakamoto.sln
@@ -19,6 +19,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "osu.Game.Rulesets.Taiko", "
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "submodules", "submodules", "{1F9350A8-2320-4911-B3E0-455AC35972B0}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sakamoto.Packet", "Sakamoto.Packet\Sakamoto.Packet.csproj", "{A31B3FC0-B5A0-42D4-B41C-16291234862F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -53,6 +55,10 @@ Global
 		{40A8FEE0-5DFA-4EA1-968D-E42E0B40EED0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{40A8FEE0-5DFA-4EA1-968D-E42E0B40EED0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{40A8FEE0-5DFA-4EA1-968D-E42E0B40EED0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A31B3FC0-B5A0-42D4-B41C-16291234862F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A31B3FC0-B5A0-42D4-B41C-16291234862F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A31B3FC0-B5A0-42D4-B41C-16291234862F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A31B3FC0-B5A0-42D4-B41C-16291234862F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
